requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.1.0
gunicorn==21.2.0
